+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This library adheres to [Semantic Versioning](https://semver.org/) and [Keep a C
 
 ## Added
 
+* Support for the `--silence` flag introduced to a few `create` commands.
 * New commands:
   * `wp bp notice` - Used to manage Sitewide notices.
   * `wp bp tool reinstall` - Alias of the `wp bp email reinstall` command, we will deprecate the latter in the future.
@@ -18,10 +19,12 @@ This library adheres to [Semantic Versioning](https://semver.org/) and [Keep a C
 * Github Action: Testing against PHP 8.3
 * Updated deprecated function from `bp_get_group_permalink` into `bp_get_group_url`
 * Activity: make tests more deterministic
-* phpDoc
+* PHPDoc improvements
+  * All `create` commands' output were standardized
   * All `delete` commands' output were standardized
 	* All `delete` commands' alias were standardized
 	* Invalid `format` option `haml` removed
+	* `post-update` command updated to remove mention of default text creation
 * Confirmation message updated for `delete` commands that accepts multiple values
 * `wp bp group invite remove` updated to `wp bp group invite uninvite` to avoid conflict with the delete/remove command